Lay groundwork for automated deletion bot
Provide a place to put our discord bot code and the credentials for
the bot when we move to testing. These are the most fundamental files
for our goals. (It may be the case that no other files are necessary!)

The structure of this code was inspired by our Nextcloud DataBot. You
can license it however you want, though you might want to bear that in
mind. But we own the copyright on that, so it shouldn't be a problem.