Experiment with binary version of Speak-To-Me DRM
This PR is mostly a meme. I wanted to tinker with the idea of running
binary (precompiled) code from our python scripts. In this example, I
compiled the code using CPython 3.9.2, so in order to test it, the
owner would presumably need to ensure his python interpreter is
compatible with that bytecode.

That is, if you trust my bytecode in the first place. Don't worry, I
promise it's safe to run.... this time. :D