Add explicit command to disconnect
In my experience, shutting down the bot is no easy feat. One needs to
send a signal (TERM or KILL) to the bot process and then wait a few
minutes for the discord server to mark the bot as offline. We've all
been there, right?

Well, this new feature solves the issue. A quick `!disconnect` and the
bot will shut down, no questions asked. Maybe we should check the
privilege of the user who issued the command, but barring these tweaks,
this is an important feature to support testing.