setup basic cogs
I've setup basic cogs importing and storage in cogs/user_commands/reply to create a baseline for importing and setting up cogs - we should port the other commands and events to cogs as well. The basic rundown is that we have a file called something.py that has a class called do_something with a command function called start. We then use a function called setup (has to be named setup) outside of the class that runs bot.add_cog(do_something(bot)). In our main function, we then can call this cog by referring to the file: await bot.load_extension('something')